requirements.txt
discord.py==2.3.2
aiohttp>=3.8.4
websockets>=11.0.2
yt-dlp>=2024.1.1
mutagen>=1.46.0
ffmpeg-python>=0.2.0
requests>=2.28.2
beautifulsoup4>=4.12.2
PyNaCl>=1.5.0
